update all, 3 ko to go

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: bajeanno <bajeanno@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 20:39:11 by bajeanno          #+#    #+#             */
-/*   Updated: 2022/11/18 02:12:55 by bajeanno         ###   ########lyon.fr   */
+/*   Updated: 2022/11/18 02:59:55 by bajeanno         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@
 void	*ft_memmove(void *dst, const void *src, size_t len);
 char	*ft_strfusion(char const *buf, int *size, char *line, int buf_size);
 char	*ft_realloc(char *str, int size);
-char	*ft_strncat(char *s1, const char *s2, unsigned int n);
+char	*ft_strncat(char *s1, const char *s2, size_t n);
 int		ft_isset(char c, char const *set);
 char	*get_next_line(int fd);
